+++ priv/repo/migrations/20240319112421_add_dataclip_gin_index.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Repo.Migrations.AddDataclipGinIndex do
+  use Ecto.Migration
+  @disable_ddl_transaction true
+  @disable_migration_lock true
+
+  def change do
+    create index(:dataclips, [:body], concurrently: true, using: :gin)
+  end
+end