+++ priv/repo/migrations/20240319125628_add_tsvector_index_to_log_lines.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Repo.Migrations.AddTsvectorIndexToLogLines do
+  use Ecto.Migration
+  @disable_ddl_transaction true
+  @disable_migration_lock true
+
+  def change do
+    create index(:log_lines, [:search_vector], using: :gin)
+  end
+end